Address minor comments
