+++ scripts/simplify_network.py
@@ -91,6 +91,7 @@ from _helpers import (
     add_year_suffix_to_carriers,
     configure_logging,
     create_logger,
+    get_linetype_by_voltage_and_country,
     nearest_shape,
     restore_base_carrier_names,
     update_config_dictionary,
@@ -108,31 +109,6 @@ from scipy.sparse.csgraph import connected_components, dijkstra
 sys.settrace
 
 logger = create_logger(__name__)
-
-
-def _get_linetype_by_voltage_and_country(v_nom, country, linetypes):
-    """
-    Return the closest available line type for a voltage and country.
-    """
-    if "default" not in linetypes:
-        raise ValueError("Missing 'default' line type mapping.")
-
-    mapping = {
-        **linetypes["default"],
-        **linetypes.get(country, {}),
-    }
-
-    if not mapping:
-        raise ValueError(
-            f"No line type mapping found for voltage {v_nom} kV "
-            f"and country '{country}'."
-        )
-
-    voltage = min(
-        mapping,
-        key=lambda candidate: abs(float(candidate) - float(v_nom)),
-    )
-    return mapping[voltage]
 
 
 def simplify_network_to_base_voltage(n, linetypes, base_voltage):
@@ -166,7 +142,7 @@ def simplify_network_to_base_voltage(n, linetypes, base_voltage):
 
     line_countries = n.lines["bus0"].map(n.buses["country"])
     n.lines["type"] = line_countries.map(
-        lambda country: _get_linetype_by_voltage_and_country(
+        lambda country: get_linetype_by_voltage_and_country(
             base_voltage,
             country,
             linetypes,
